Add :app:ui:base module and move K9Activity to it
