+++ app/k9mail-jmap/src/main/java/com/fsck/k9/App.kt
@@ -3,7 +3,7 @@ package com.fsck.k9
 import android.app.Application
 import com.fsck.k9.activity.MessageCompose
 import com.fsck.k9.controller.MessagingController
-import com.fsck.k9.ui.ThemeManager
+import com.fsck.k9.ui.base.ThemeManager
 import org.koin.android.ext.android.inject
 
 class App : Application() {
