+++ app/k9mail-jmap/src/main/java/com/fsck/k9/ui/addaccount/AddAccountActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.navigation.NavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
-import com.fsck.k9.activity.K9Activity
 import com.fsck.k9.jmap.R
+import com.fsck.k9.ui.base.K9Activity
 import com.fsck.k9.ui.findNavController
 
 class AddAccountActivity : K9Activity() {
