+++ app/k9mail/src/main/java/com/fsck/k9/widget/unread/UnreadWidgetConfigurationActivity.kt
@@ -3,7 +3,7 @@ package com.fsck.k9.widget.unread
 import android.appwidget.AppWidgetManager
 import android.os.Bundle
 import com.fsck.k9.R
-import com.fsck.k9.activity.K9Activity
+import com.fsck.k9.ui.base.K9Activity
 import com.fsck.k9.ui.fragmentTransaction
 import timber.log.Timber
