+++ app/ui/base/src/main/java/com/fsck/k9/ui/base/K9Activity.kt
@@ -1,11 +1,9 @@
-package com.fsck.k9.activity
+package com.fsck.k9.ui.base
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.fsck.k9.ui.R
-import com.fsck.k9.ui.ThemeManager
 
 abstract class K9Activity : AppCompatActivity() {
     private val base = K9ActivityCommon(this, ThemeType.DEFAULT)
