+++ app/ui/base/src/main/java/com/fsck/k9/ui/base/KoinModule.kt
@@ -1,0 +1,7 @@
+package com.fsck.k9.ui.base
+
+import org.koin.dsl.module
+
+val uiBaseModule = module {
+    single { ThemeManager(context = get(), themeProvider = get()) }
+}
