+++ app/ui/base/src/main/java/com/fsck/k9/ui/base/ThemeProvider.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.ui
+package com.fsck.k9.ui.base
 
 import androidx.annotation.StyleRes
