+++ app/ui/legacy/src/main/java/com/fsck/k9/UiKoinModules.kt
@@ -5,6 +5,7 @@ import com.fsck.k9.activity.activityModule
 import com.fsck.k9.autodiscovery.providersxml.autodiscoveryProvidersXmlModule
 import com.fsck.k9.contacts.contactsModule
 import com.fsck.k9.fragment.fragmentModule
+import com.fsck.k9.ui.base.uiBaseModule
 import com.fsck.k9.ui.choosefolder.chooseFolderUiModule
 import com.fsck.k9.ui.endtoend.endToEndUiModule
 import com.fsck.k9.ui.folders.foldersUiModule
@@ -15,17 +16,18 @@ import com.fsck.k9.ui.uiModule
 import com.fsck.k9.view.viewModule
 
 val uiModules = listOf(
-        activityModule,
-        uiModule,
-        settingsUiModule,
-        endToEndUiModule,
-        foldersUiModule,
-        messageListUiModule,
-        manageFoldersUiModule,
-        chooseFolderUiModule,
-        fragmentModule,
-        contactsModule,
-        accountModule,
-        autodiscoveryProvidersXmlModule,
-        viewModule
+    uiBaseModule,
+    activityModule,
+    uiModule,
+    settingsUiModule,
+    endToEndUiModule,
+    foldersUiModule,
+    messageListUiModule,
+    manageFoldersUiModule,
+    chooseFolderUiModule,
+    fragmentModule,
+    contactsModule,
+    accountModule,
+    autodiscoveryProvidersXmlModule,
+    viewModule
 )
