+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/EditIdentity.kt
@@ -6,6 +6,7 @@ import com.fsck.k9.Account
 import com.fsck.k9.Identity
 import com.fsck.k9.Preferences
 import com.fsck.k9.ui.R
+import com.fsck.k9.ui.base.K9Activity
 import kotlinx.android.synthetic.main.edit_identity.*
 
 class EditIdentity : K9Activity() {
