+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/K9ListActivity.java
@@ -8,6 +8,7 @@ import android.widget.ListAdapter;
 import android.widget.ListView;
 
 import com.fsck.k9.K9;
+import com.fsck.k9.ui.base.K9Activity;
 
 
 public abstract class K9ListActivity extends K9Activity {
