+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/MessageCompose.java
@@ -105,7 +105,8 @@ import com.fsck.k9.message.SimpleMessageFormat;
 import com.fsck.k9.search.LocalSearch;
 import com.fsck.k9.ui.EolConvertingEditText;
 import com.fsck.k9.ui.R;
-import com.fsck.k9.ui.ThemeManager;
+import com.fsck.k9.ui.base.K9Activity;
+import com.fsck.k9.ui.base.ThemeManager;
 import com.fsck.k9.ui.compose.QuotedMessageMvpView;
 import com.fsck.k9.ui.compose.QuotedMessagePresenter;
 import com.fsck.k9.ui.helper.SizeFormatter;
