+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/MessageList.java
@@ -53,7 +53,8 @@ import com.fsck.k9.search.SearchSpecification.SearchField;
 import com.fsck.k9.ui.BuildConfig;
 import com.fsck.k9.ui.K9Drawer;
 import com.fsck.k9.ui.R;
-import com.fsck.k9.ui.Theme;
+import com.fsck.k9.ui.base.K9Activity;
+import com.fsck.k9.ui.base.Theme;
 import com.fsck.k9.ui.managefolders.ManageFoldersActivity;
 import com.fsck.k9.ui.messagelist.DefaultFolderProvider;
 import com.fsck.k9.ui.messageview.MessageViewFragment;
