+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/setup/AccountSetupComposition.java
@@ -12,7 +12,7 @@ import android.widget.RadioButton;
 import com.fsck.k9.Account;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.ui.R;
-import com.fsck.k9.activity.K9Activity;
+import com.fsck.k9.ui.base.K9Activity;
 
 public class AccountSetupComposition extends K9Activity {
