+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/setup/AccountSetupOptions.java
@@ -15,7 +15,7 @@ import com.fsck.k9.Account;
 import com.fsck.k9.Core;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.ui.R;
-import com.fsck.k9.activity.K9Activity;
+import com.fsck.k9.ui.base.K9Activity;
 
 
 public class AccountSetupOptions extends K9Activity implements OnClickListener {
