+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/KoinModule.kt
@@ -1,6 +1,7 @@
 package com.fsck.k9.ui
 
 import android.content.Context
+import com.fsck.k9.ui.base.ThemeProvider
 import com.fsck.k9.ui.helper.DisplayHtmlUiFactory
 import com.fsck.k9.ui.helper.HtmlSettingsProvider
 import com.fsck.k9.ui.helper.HtmlToSpanned
@@ -10,7 +11,6 @@ import org.koin.dsl.module
 val uiModule = module {
     single { HtmlToSpanned() }
     single<ThemeProvider> { K9ThemeProvider() }
-    single { ThemeManager(context = get(), themeProvider = get()) }
     single { HtmlSettingsProvider(get()) }
     single { DisplayHtmlUiFactory(get()) }
     factory { (context: Context) -> SizeFormatter(context.resources) }
