+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/helper/HtmlSettingsProvider.kt
@@ -2,8 +2,8 @@ package com.fsck.k9.ui.helper
 
 import com.fsck.k9.K9
 import com.fsck.k9.message.html.HtmlSettings
-import com.fsck.k9.ui.Theme
-import com.fsck.k9.ui.ThemeManager
+import com.fsck.k9.ui.base.Theme
+import com.fsck.k9.ui.base.ThemeManager
 
 class HtmlSettingsProvider(private val themeManager: ThemeManager) {
     fun createForMessageView() = HtmlSettings(
